SymbolicInference

Documentation for SymbolicInference.

SymbolicInference.extract_recurrences

SymbolicInference.extract_recurrences_cross

SymbolicInference.plot_motifs

SymbolicInference.plot_motifs_cross

SymbolicInference.rec_matrix_motifs

SymbolicInference.rec_matrix_probs

SymbolicInference.extract_recurrences — Method

extractrecurrences(datasource::Vector{Float64}, motifsdict::Dict{String, Vector}; numwindows::Int64 = 3)

This function returns x and y coordinates for a given window considering start and size of each motif. The y coordinates are the values from data provided by the user.

SymbolicInference.extract_recurrences_cross — Method

extract_recurrences_cross(data_source::Vector{Float64}, data_source2::Vector{Float64},
    motifs_dict::Dict{String, Vector}; num_windows::Int64 = 3)

This function returns x and y coordinates for a given window considering start and size of each motif. The y coordinates are the values from data provided by the user.

SymbolicInference.plot_motifs — Method

plot_motifs(time_series::Vector{Float64},coordinates::Vector{Any}; plot_size=(2000, 1000), n_motifs=2)

Plot motifs

SymbolicInference.plot_motifs_cross — Method

plot_motifs_cross(time_series::Vector{Float64},time_series2::Vector{Float64},
    coordinates::Vector{Any}; plot_size=(2000, 1000), n_motifs=2)

Plot motifs

SymbolicInference.rec_matrix_motifs — Method

rec_matrix_motifs(rec_matrix::RecurrenceMatrix;seqs="recurrences",max_window=6,n_motifs=2)

Returns set of probabilities associated with consecutive runs in off-diagonals.

Argument seqs sets the type of consecutive sequences: either 'double' (recurrences and non-recurrences), 'recurrences' or 'poincare' (non-recurrences). The first n diagonals (given by max_window argument) are considered, along with nmotifs for each diagonal. See `AnalyticComb.weightedbinrunsprob` for definition of symbolic construction.

SymbolicInference.rec_matrix_probs — Method

rec_matrix_probs(rec_matrix::RecurrenceMatrix;seqs="double",max_window=6)

Returns set of probabilities associated with consecutive runs in off-diagonals.

Argument seqs sets the type of consecutive sequences: either 'double' (recurrences and non-recurrences), 'recurrences' or 'poincare' (non-recurrences). The first n diagonals (given by max_window argument) are considered. See also AnalyticComb.p_val_weighted and AnalyticComb.weighted_bin_runs_coeff.