SymbolicInference

Documentation for SymbolicInference.

SymbolicInference.extract_recurrences

SymbolicInference.extract_recurrences_cross

SymbolicInference.extract_recurrences_joint

SymbolicInference.persistence_barcode

SymbolicInference.persistence_barcode

SymbolicInference.plot_motifs

SymbolicInference.plot_motifs_cross

SymbolicInference.plot_motifs_joint

SymbolicInference.rec_matrix_motifs

SymbolicInference.extract_recurrences — Method

extractrecurrences(datasource::Vector{Float64}, motifsdict::Dict{String, Vector}; numwindows::Int64 = 3)

This function returns x and y coordinates for a given window considering start and size of each motif. The y coordinates are the values from data provided by the user.

SymbolicInference.extract_recurrences_cross — Method

extract_recurrences_cross(data_source::Vector{Float64}, data_source2::Vector{Float64},
    motifs_dict::Dict{String, Vector}; num_windows::Int64 = 3)

This function returns x and y coordinates for a given window considering start and size of each motif detected from two time-series in cross-recurrence matrices . The y coordinates are the values from data provided by the user.

SymbolicInference.extract_recurrences_joint — Method

extract_recurrences_joint(data_source::Vector{Float64}, data_source2::Vector{Float64},
    motifs_dict::Dict{String, Vector}; num_windows::Int64 = 3)

    This function returns x and y coordinates for a given window 
        considering start and size of each motif detected from 
        two time-series in joint-recurrence matrices .

The y coordinates are the values from data provided by the user.

SymbolicInference.persistence_barcode — Method

persistencebarcode(timeseries1,timeseries2; range = collect(0.1:0.1:0.9), nwindows=10,alpha_thresh=0.05)

Return barcode plot for each window. Points are plotted whenever the p-value is smaller than alpha_thresh.

SymbolicInference.persistence_barcode — Method

persistencebarcode(timeseries; range = collect(0.1:0.1:0.9), nwindows=10,alphathresh=0.05)

Return barcode plot for each window. Points are plotted whenever the p-value is smaller than alpha_thresh.

SymbolicInference.plot_motifs — Method

plot_motifs(time_series::Vector{Float64},coordinates::Vector{Any}; plot_size=(2000, 1000), n_motifs=2)

Plot motifs from coordinates extracted with extract_recurrences.

SymbolicInference.plot_motifs_cross — Method

plot_motifs_cross(time_series::Vector{Float64},time_series2::Vector{Float64},
    coordinates::Vector{Any}; plot_size=(2000, 1000), n_motifs=2)

Plot motifs from coordinates extracted with extract_recurrences_cross.

SymbolicInference.plot_motifs_joint — Method

plot_motifs_joint(time_series::Vector{Float64},time_series2::Vector{Float64},
    coordinates::Vector{Any}; plot_size=(2000, 1000), n_motifs=2)

Plot motifs from coordinates extracted with extract_recurrences_joint.

SymbolicInference.rec_matrix_motifs — Method

rec_matrix_motifs(rec_matrix::RecurrenceMatrix;seqs="recurrences",window_range=collect(1:6),n_motifs=2)

Returns set of probabilities associated with consecutive runs in off-diagonals.

Argument seqs sets the type of consecutive sequences: either 'double' (recurrences and non-recurrences), 'recurrences' or 'poincare' (non-recurrences). The diagonals given by window_range argument are considered, along with nmotifs for each diagonal. See `AnalyticComb.weightedbinrunsprob` for definition of symbolic construction.